Gradient Option, Masthead Images
Gradients can now be inputted by the user, they have a text area to
write the input values in. Can also turn off the block if they want BG
images or solid color. Added jquery to detect an empty Masthead Image
slot and remove it from the array while outputting to randomization.
So, if all 5 aren't full it'll clear the empty ones.